Fixed bugs

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoHibernateImpl.java
@@ -6,6 +6,7 @@ import org.hibernate.Session;
 import org.hibernate.Transaction;
 import org.hibernate.query.Query;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class UserDaoHibernateImpl implements UserDao {
@@ -19,10 +20,7 @@ public class UserDaoHibernateImpl implements UserDao {
             session.createSQLQuery(sql).executeUpdate();
             transaction.commit();
         } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            e.printStackTrace();
+            throw new RuntimeException(e);
         }
     }
 
@@ -33,11 +31,9 @@ public class UserDaoHibernateImpl implements UserDao {
         try (Session session = Util.getSessionFactory().openSession()) {
             transaction = session.beginTransaction();
             session.createSQLQuery(sql).executeUpdate();
-
+            transaction.commit();
         } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
+            throw new RuntimeException(e);
         }
 
     }
@@ -64,7 +60,9 @@ public class UserDaoHibernateImpl implements UserDao {
         try (Session session = Util.getSessionFactory().openSession()) {
             transaction = session.beginTransaction();
             User user = session.get(User.class, id);
-            session.delete(user);
+            if (user != null) {
+                session.delete(user);
+            }
             transaction.commit();
         } catch (Exception e) {
             if (transaction != null) {
@@ -76,7 +74,7 @@ public class UserDaoHibernateImpl implements UserDao {
     @Override
     public List<User> getAllUsers() {
         Transaction transaction = null;
-        List<User> users = null;
+        List<User> users = new ArrayList<>();
         try (Session session = Util.getSessionFactory().openSession()) {
             transaction = session.beginTransaction();
             Query<User> query = session.createQuery("FROM User", User.class);
@@ -99,10 +97,9 @@ public class UserDaoHibernateImpl implements UserDao {
             transaction = session.beginTransaction();
             Query query = session.createSQLQuery(sql);
             query.executeUpdate();
+            transaction.commit();
         } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
+            throw new RuntimeException(e);
         }
     }
 }
